requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
httpx==0.28.1
google-genai==1.65.0
python-dotenv==1.2.2
